Making it functional.
Thus it is a functional plugin. The importer, once activated, will walk backwards until the first checkin. It'll also check for new checkins.

Featured images are the most recent check-in image, if available, or the most recent badge awarded, if no check-in image was submitted with the latest checkin.

I have no idea what will happen on the importer when it hits the first checkin. It should remove the option alerting the plugin we're in Import mode.

There is a ton of code standard issues. Code needs to be moved around. Modularized, and whatnot. But alas, this needed to be shipped today, so here it is. 0.1.0-so-alpha-it-hurts.